+++ inject/apple.js
@@ -1,6 +1,6 @@
 /* global contentScopeFeatures */
 
-import { processConfig } from './../src/apple-utils'
+import { processConfig } from './../src/utils'
 
 function init () {
     const processedConfig = processConfig($CONTENT_SCOPE$, $USER_UNPROTECTED_DOMAINS$, $USER_PREFERENCES$)
