+++ scripts/generateSJCL.js
@@ -10,11 +10,7 @@ async function init () {
         return
     }
 
-    const command = process.platform === 'win32'
-        ? 'cd node_modules/sjcl/ && perl ./configure --no-export --compress=none --without-all --with-hmac --with-codecHex && make'
-        : 'cd node_modules/sjcl/ && ./configure --no-export --compress=none --without-all --with-hmac --with-codecHex && make'
-    await exec(command)
-
+    await exec('cd node_modules/sjcl/ && perl ./configure --no-export --compress=none --without-all --with-hmac --with-codecHex && make')
     const sjclFileContents = await fs.readFile('node_modules/sjcl/sjcl.js')
     // Reexport the file as es6 module format
     const contents = `// @ts-nocheck
